feat: core 모듈 리펙토링
