+++ android/app/build.gradle.kts
@@ -33,6 +33,7 @@ android {
 }
 
 dependencies {
+    implementation(project(path = ":network"))
     implementation(project(path = ":core"))
     implementation(project(path = ":shared"))
     implementation(project(path = ":features:core-ui"))
