+++ android/core/src/main/java/com/hmju/core/login_manager/di/LoginManagerModule.kt
@@ -1,0 +1,22 @@
+package com.hmju.core.login_manager.di
+
+import com.hmju.core.login_manager.LoginManager
+import com.hmju.core.login_manager.LoginManagerImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+/**
+ * Description : 로그인 메니저 모듈
+ *
+ * Created by juhongmin on 2022/01/12
+ */
+@InstallIn(SingletonComponent::class)
+@Module
+internal abstract class LoginManagerModule {
+    @Singleton
+    @Binds
+    abstract fun bindLoginManager(loginManager: LoginManagerImpl): LoginManager
+}
