+++ android/core/src/main/java/com/hmju/core/rxbus/SimpleLikeEvent.kt
@@ -1,0 +1,11 @@
+package com.hmju.core.rxbus
+
+/**
+ * Description :
+ *
+ * Created by juhongmin on 2022/08/23
+ */
+data class SimpleLikeEvent(
+    val isAdd: Boolean,
+    val id: Long
+) : RxBusEvent()
