+++ android/features/base-mvvm/src/main/java/com/features/base_mvvm/ui/mvvm_lifecycle/MvvmLifecycleTest2ViewModel.kt
@@ -9,7 +9,7 @@ import com.features.core_ui.lifecycle.OnActivityResult
 import com.features.core_ui.lifecycle.OnCreated
 import com.features.core_ui.lifecycle.OnResumed
 import com.hmju.core.data.model.params.GoodsParamMap
-import com.hmju.shared.login_manager.LoginManager
+import com.hmju.core.login_manager.LoginManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import timber.log.Timber
