+++ android/features/base-mvvm/src/main/java/com/features/base_mvvm/ui/mvvm_lifecycle/MvvmLifecycleTestViewModel.kt
@@ -11,7 +11,7 @@ import com.features.core_ui.base.BaseActivity
 import com.features.core_ui.base.IntentKey
 import com.features.core_ui.lifecycle.*
 import com.hmju.core.data.model.params.GoodsParamMap
-import com.hmju.shared.login_manager.LoginManager
+import com.hmju.core.login_manager.LoginManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.kotlin.addTo
 import io.reactivex.rxjava3.schedulers.Schedulers
