+++ android/features/core-ui/src/main/java/com/features/core_ui/binding/CommonBindingAdapter.kt
@@ -4,7 +4,7 @@ import android.view.View
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.text.HtmlCompat
 import androidx.databinding.BindingAdapter
-import com.hmju.shared.like_manager.LikeManager
+import com.hmju.core.like_manager.LikeManager
 
 /**
  * Description :
