+++ android/features/main/build.gradle.kts
@@ -13,6 +13,7 @@ android {
 
 dependencies {
     implementation(project(":shared"))
+    implementation(project(":core"))
     implementation(project(":features:core-ui"))
     implementation(project(":features:recyclerview-requirements"))
     implementation(project(":features:network-requirements"))
