+++ android/features/main/src/main/java/com/features/main/MainViewModel.kt
@@ -6,7 +6,7 @@ import com.features.core_ui.lifecycle.OnCreated
 import com.features.core_ui.lifecycle.OnIntent
 import com.features.network_requirements.NetworkRequirements
 import com.features.recyclerview_requirements.RecyclerViewRequirements
-import com.hmju.shared.login_manager.LoginManager
+import com.hmju.core.login_manager.LoginManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.core.Flowable
 import io.reactivex.rxjava3.kotlin.addTo
