+++ android/features/network/src/main/java/com/features/network/ui/expired_token/RefreshTokenFragment.kt
@@ -8,7 +8,7 @@ import com.features.network.BR
 import com.features.network.R
 import com.features.network.databinding.FRefreshTokenBinding
 import com.features.network.usecase.*
-import com.hmju.shared.login_manager.LoginManager
+import com.hmju.core.login_manager.LoginManager
 import dagger.hilt.android.AndroidEntryPoint
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.core.Flowable
