+++ android/network/build.gradle.kts
@@ -1,0 +1,49 @@
+plugins {
+    id("com.android.library")
+    id("kotlinx-serialization")
+    id("dagger.hilt.android.plugin")
+    kotlin("android")
+    kotlin("kapt")
+}
+
+android {}
+
+dependencies {
+
+    implementation(project(":core"))
+    implementation(project(":shared"))
+
+    /**
+     * Network
+     */
+    implementation(Retrofit.base)
+    implementation(Retrofit.okhttp)
+    implementation(Retrofit.rxjava)
+    implementation(Retrofit.kotlinx)
+    implementation(Retrofit.okhttpLogger)
+
+    /**
+     * Kotlinx Serialization
+     */
+    implementation(KotlinX.serialization)
+
+    /**
+     * Rx
+     */
+    implementation(Rx.java)
+    implementation(Rx.kotlin)
+
+    /**
+     * Hilt
+     */
+    implementation(Hilt.android)
+    kapt(Hilt.compiler)
+
+    /**
+     * Unit Test
+     */
+    testImplementation(UnitTest.junit)
+    androidTestImplementation(UnitTest.junit)
+    androidTestImplementation(UnitTest.androidJUnit)
+    androidTestImplementation(UnitTest.Espresso.core)
+}
