+++ android/network/src/main/java/com/hmju/network/NetworkConfig.kt
@@ -1,4 +1,4 @@
-package com.hmju.core.data
+package com.hmju.network
 
 /**
  * Description : Network 설정 정의 클래스
@@ -6,6 +6,9 @@ package com.hmju.core.data
  * Created by juhongmin on 2022/01/12
  */
 object NetworkConfig {
+    const val CONNECT_TIME_OUT = 5000L
+    const val READ_TIME_OUT = 5000L
+    const val WRITE_TIME_OUT = 5000L
     const val HEADER_KEY_ACCEPT = "accept"
     const val HEADER_VAL_ACCEPT = "application/json"
     const val HEADER_KEY_CONTENT = "Content-Type"
