+++ android/network/src/main/java/com/hmju/network/impl/AuthRepositoryImpl.kt
@@ -1,0 +1,26 @@
+package com.hmju.network.impl
+
+import com.hmju.core.data.model.auth.TokenEntity
+import com.hmju.core.data.model.base.JSendObj
+import com.hmju.network.AuthApiService
+import com.hmju.core.data.repository.AuthRepository
+import io.reactivex.rxjava3.core.Single
+import javax.inject.Inject
+
+/**
+ * Description : Auth Repository 구현체 클래스
+ *
+ * Created by juhongmin on 2022/01/12
+ */
+internal class AuthRepositoryImpl @Inject constructor(
+    private val authApiService: AuthApiService
+) : AuthRepository {
+
+    override fun tokenRefresh(): Single<JSendObj<TokenEntity>> {
+        return authApiService.tokenRefresh()
+    }
+
+    override fun tokenExpired(): Single<JSendObj<TokenEntity>> {
+        return authApiService.tokenExpired()
+    }
+}
