+++ android/settings.gradle.kts
@@ -15,3 +15,4 @@ include(":features:main")
 include(":features:base-mvvm")
 include(":features:base-mvvm-requirements")
 include(":test")
+include(":network")
