--- android/shared/src/main/java/com/hmju/shared/rxbus/SimpleLikeEvent.kt
@@ -1,11 +1,0 @@
-package com.hmju.shared.rxbus
-
-/**
- * Description :
- *
- * Created by juhongmin on 2022/08/23
- */
-data class SimpleLikeEvent(
-    val isAdd: Boolean,
-    val id: Long
-) : RxBusEvent()
